job/requirements.txt
cloud-sql-python-connector~=0.8.1
pg8000~=1.29.1
SQLAlchemy~=1.4.41
json-logging==1.3.0
pandas~=1.5.0
google-cloud-storage~=2.5.0